move activity type selectors into seperate files
closes #404
