+++ projects/ngx-d3timeline/src/lib/activity/activity.selectors.ts
@@ -3,7 +3,7 @@ import {
   selectPositionedActivities,
   selectLastDraggedActivityId
 } from '../store/state';
-import { selectTimeOrientation } from '../options/options.selectors';
+import { selectTimeOrientation } from '../options/selectors/options.selectors';
 import { createSelector } from '../store-lib/selector/create-selector';
 import {
   findActivity,
