+++ projects/ngx-d3timeline/src/lib/resource-rectangle/resource-rectangle.selectors.ts
@@ -6,7 +6,7 @@ import {
 import { selectView } from '../store/state';
 import { getResourceRectangles } from './resource-rectangle-utils';
 import { selectViewClipRectangleHeight } from '../view/view.selectors';
-import { selectResourceAxisFontSize } from '../options/axis-options.selectors';
+import { selectResourceAxisFontSize } from '../options/selectors/axis-options.selectors';
 
 // This could be better, tidy up as part of #373
 export const selectResourceRectangles = createSelector(
