+++ projects/ngx-d3timeline/src/lib/scales/scale-selectors.ts
@@ -8,8 +8,8 @@ import { configureBandScale, rescaleTime } from './scale-utils';
 import {
   selectResourceOrientation,
   selectTimeOrientation
-} from '../options/options.selectors';
-import { selectResourceGap } from '../options/resource-options.selectors';
+} from '../options/selectors/options.selectors';
+import { selectResourceGap } from '../options/selectors/resource-options.selectors';
 import { getOrientedScale } from './oriented-scale';
 
 export const selectBandScale = createSelector(
